chore: cleanup validation schema for properties
- add rule to rate setting, cannot exceed 100 if percentage
